# triedb/pathdb: improve perf by separating nodes map