@rjl493456442 2025-03-05
<img width="1421" alt="截屏2025-03-05 12 44 48" src="https://github.com/user-attachments/assets/dbf10580-7875-4995-a2d3-c3a33a4e500e" />
<img width="1401" alt="截屏2025-03-05 12 45 00" src="https://github.com/user-attachments/assets/42e48208-4db6-434b-a89e-424bb857d881" />
<img width="1422" alt="截屏2025-03-05 12 46 22" src="https://github.com/user-attachments/assets/fb7b5af1-2c2d-449e-881f-7afa8287454c" />

This PR (bench07) is slightly faster than master, due to a 1.8ms reduction in account trie update time (trie node loading).